Use uint8_t for 'bpp' (bits per pixel)

+++ include/render.h
@@ -231,7 +231,7 @@ std::deque<std::string> RENDER_InventoryShaders();
 void RENDER_SetSize(const uint16_t width, const uint16_t height,
                     const bool double_width, const bool double_height,
                     const Fraction& render_pixel_aspect_ratio,
-                    const uint16_t bits_per_pixel,
+                    const uint8_t bits_per_pixel,
                     const double frames_per_second, const VideoMode& video_mode);
 
 bool RENDER_StartUpdate(void);
